correción de examen

+++ src/edu/uaslp/examen/Book.java
@@ -1,33 +1,41 @@
 package edu.uaslp.examen;
 
 public class Book {
+
     private String title;
     private String isbn;
     private String author;
-    private short stars;
+    private int stars;
 
     public String getTitle() {
         return title;
     }
+
     public void setTitle(String title) {
         this.title = title;
     }
+
     public String getIsbn() {
         return isbn;
     }
+
     public void setIsbn(String isbn) {
         this.isbn = isbn;
     }
+
     public String getAuthor() {
         return author;
     }
+
     public void setAuthor(String author) {
         this.author = author;
     }
-    public short getStars() {
+
+    public int getStars() {
         return stars;
     }
-    public void setStars(short stars) {
+
+    public void setStars(int stars) {
         this.stars = stars;
     }
 }
